+++ src/Controller/Admin/OrderArticle.php
@@ -9,7 +9,7 @@ namespace OxidSolutionCatalysts\AmazonPay\Controller\Admin;
 
 use OxidEsales\Eshop\Core\Exception\DatabaseConnectionException;
 use OxidEsales\Eshop\Core\Exception\DatabaseErrorException;
-use OxidEsales\EshopCommunity\Core\Request;
+use OxidEsales\Eshop\Core\Registry;
 use OxidSolutionCatalysts\AmazonPay\Core\Config;
 use OxidSolutionCatalysts\AmazonPay\Core\Constants;
 use OxidSolutionCatalysts\AmazonPay\Core\Logger;
@@ -56,10 +56,9 @@ class OrderArticle extends OrderArticle_parent
         if (!$config->automatedRefundActivated()) {
             return;
         }
-        $request = new Request();
         // get article id
         /** @var string $sOrderArtId */
-        $sOrderArtId = $request->getRequestParameter('sArtID') ?: '';
+        $sOrderArtId = Registry::getRequest()->getRequestParameter('sArtID') ?: '';
         $sOrderId = $this->getEditObjectId() ?: '';
 
         $oOrderArticle = oxNew(\OxidEsales\Eshop\Application\Model\OrderArticle::class);